Boundary structures were refactored. Boundaries with polar coordinates were added
